requirements.txt
Babel==2.6.0
beautifulsoup4==4.8.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Django==2.2.1
django-audit-log==0.7.0
django-cors-headers==3.0.1
django-cors-middleware==1.4.0
django-discover-runner==1.0
django-fcm==0.1.1
django-filter==2.1.0
django-oauth-toolkit==1.2.0
django-phone-field==1.7.1
django-phonenumber-field==2.4.0
django-rest-auth==0.9.5
django-user-tracking==0.4.11
django-uuidfield==0.5.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
drf-tracking==1.5.0
fcm-django==0.2.21
gunicorn==19.9.0
idna==2.8
Jinja2==2.10.1
livereload==2.6.1
Markdown==3.1.1
MarkupSafe==1.1.1
mkdocs==1.0.4
oauthlib==3.0.2
phonenumbers==8.10.12
Pillow==6.0.0
pkg-resources==0.0.0
psycopg2==2.8.2
psycopg2-binary==2.8.2
pyfcm==1.4.5
PyJWT==1.7.1
python-decouple==3.1
python-memcached==1.59
pytz==2019.1
PyYAML==5.1
requests==2.21.0
requests-toolbelt==0.9.1
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
tornado==6.0.2
urllib3==1.24.3
